// src/app/claim-processing/manual-claim-processing/process-claim-home-health/model/screen-save-area.model.ts
import {ScreenChargeLine} from './screen-charge-line.model';

/**
 * Model class ScreenSaveArea
 * Path: screenbean/procclmhhcoohfgncov
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmhhcoohfgncov::ScreenSaveArea
 * Legacy Mapping: SCREEN-SAVE-AREA
 */
export class ScreenSaveArea {
  filler1 = '';
  mapCommandLine = '';
  filler2 = '';
  screenIcdCodes = '';
  filler6 = '';
  screenChargeLines: ScreenChargeLine[] = [];
  screenExPay1 = '';
  filler15 = '';
  screenExPay2 = '';
  filler16 = '';
  screenAfterTerm = '';
  m28err1 = '';
  m28lne2 = '';
  m28err2 = '';
  m28atrm = '';
  m28xpa2 = '';
  m28head = '';
  m28xpa1 = '';
  m28tsdt = '';
  m28rel2 = '';
  m28rel1 = '';
  m28lne1 = '';
  m28tlt1 = '';
  m28cno6 = '';
  m28icd2 = '';
  m28dob = '';
  m28assc = '';
  m28cno5 = '';
  m28mem2 = '';
  m28cno3 = '';
  m28cno4 = '';
  m28cno2 = '';
  m28icd3 = '';
  m28com2 = '';
  m28trm2 = '';
  m28pln2 = '';
  m28eff2 = '';
  m28cno1 = '';
  m28name = '';
  m28pln1 = '';
  m28eff3 = '';
  m28mem1 = '';
  m28memn = '';
  m28pln3 = '';
  m28com1 = '';
  m28eff1 = '';
  m28icd1 = '';
  m28mem3 = '';
  m28cmnd = '';
  m28trm1 = '';
  m28trm3 = '';
}